approximate output type changed

+++ inc/display2.h
@@ -102,7 +102,7 @@ uint_fast16_t normalize(
 	uint_fast16_t range		// включает выходное значение
 	);
 
-uint_fast16_t approximate(
+int_fast32_t approximate(
 	const int16_t * points,		// массив позиций входных значений
 	const int_fast32_t * angles,		// массив позицый выходных значений
 	unsigned n,					// размерность массивов
